+++ CHANGELOG.rst
@@ -7,6 +7,16 @@ The format is based on `Keep a Changelog
 <https://keepachangelog.com/en/1.0.0/>`_, and this project adheres to `Semantic
 Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
+0.7.5
+-----
+
+Changed:
+
+- Changed master dataset for first card when marker channels are enabled to
+  the second real channel instead of the marker channel. This number is then
+  more comparable to the counters for the other frame processors.
+
+
 0.7.4
 -----